+++ codemods/misc/migrate-mui-chip-to-tag/tests/aliased-import/expected.tsx
@@ -1,0 +1,5 @@
+import { Tag } from '@backstage/ui';
+
+const MyComponent = () => (
+  <Tag size="small">Tag</Tag>
+);