+++ codemods/misc/migrate-mui-dialog-to-bui-dialog/tests/partial-barrel-import/expected.tsx
@@ -1,0 +1,9 @@
+import { Button } from '@material-ui/core';
+import { Dialog, DialogBody, DialogHeader } from '@backstage/ui';
+
+const MyDialog = ({ open, onClose }: { open: boolean; onClose: () => void }) => (
+  <>
+    <Dialog isOpen={open} onOpenChange={isOpen => !isOpen && onClose()}><DialogHeader>Info</DialogHeader><DialogBody>Details here</DialogBody></Dialog>
+    <Button>Keep me</Button>
+  </>
+);