+++ codemods/misc/migrate-mui-list-family-to-bui-list/tests/interactive-todo/expected.tsx
@@ -4,9 +4,11 @@ import ListItemText from '@material-ui/core/ListItemText';
 
 const MyComponent = () => (
   <List>
-    <>{/* TODO(backstage-codemod): verify nonstandard list row manually */}
-<ListItem button onClick={handleClick}>
+    <>
+  {/* TODO(backstage-codemod): verify nonstandard list row manually */}
+  <ListItem button onClick={handleClick}>
       <ListItemText primary="Clickable" />
-    </ListItem></>
+    </ListItem>
+</>
   </List>
 );