+++ codemods/misc/migrate-mui-menu-popover-to-bui-menu/tests/anchor-el-todo/expected.tsx
@@ -2,8 +2,10 @@ import Menu from '@material-ui/core/Menu';
 import MenuItem from '@material-ui/core/MenuItem';
 
 const MyComponent = ({ anchorEl, open, onClose }: any) => (
-  <>{/* TODO(backstage-codemod): finish menu host migration manually (anchorEl) */}
-<Menu anchorEl={anchorEl} open={open} onClose={onClose}>
+  <>
+  {/* TODO(backstage-codemod): finish menu host migration manually (anchorEl) */}
+  <Menu anchorEl={anchorEl} open={open} onClose={onClose}>
     <MenuItem onClick={handleAction}>Action</MenuItem>
-  </Menu></>
+  </Menu>
+</>
 );