+++ codemods/misc/migrate-mui-menu-popover-to-bui-menu/tests/named-barrel-import/expected.tsx
@@ -1,8 +1,10 @@
 import { Menu, MenuItem } from '@material-ui/core';
 
 const MyComponent = () => (
-  <>{/* TODO(backstage-codemod): finish menu host migration manually (no-trigger-element) */}
-<Menu open={isOpen}>
+  <>
+  {/* TODO(backstage-codemod): finish menu host migration manually (no-trigger-element) */}
+  <Menu open={isOpen}>
     <MenuItem onClick={doStuff}>Do stuff</MenuItem>
-  </Menu></>
+  </Menu>
+</>
 );