+++ codemods/misc/migrate-mui-menu-popover-to-bui-menu/tests/popover-with-menu-list/expected.tsx
@@ -3,10 +3,12 @@ import MenuList from '@material-ui/core/MenuList';
 import MenuItem from '@material-ui/core/MenuItem';
 
 const MyComponent = ({ open, onClose }: { open: boolean; onClose: () => void }) => (
-  <>{/* TODO(backstage-codemod): finish menu host migration manually (no-trigger-element) */}
-<Popover open={open} onClose={onClose}>
+  <>
+  {/* TODO(backstage-codemod): finish menu host migration manually (no-trigger-element) */}
+  <Popover open={open} onClose={onClose}>
     <MenuList>
       <MenuItem onClick={handleAction}>Action</MenuItem>
     </MenuList>
-  </Popover></>
+  </Popover>
+</>
 );