+++ codemods/misc/migrate-mui-tabs-to-bui-tabs/tests/vertical-tabs-todo/expected.tsx
@@ -2,9 +2,11 @@ import Tabs from '@material-ui/core/Tabs';
 import Tab from '@material-ui/core/Tab';
 
 const MyComponent = () => (
-  <>{/* TODO(backstage-codemod): verify custom tab orientation or selection logic manually */}
-<Tabs value={selected} orientation="vertical" onChange={handleChange}>
+  <>
+  {/* TODO(backstage-codemod): verify custom tab orientation or selection logic manually */}
+  <Tabs value={selected} orientation="vertical" onChange={handleChange}>
     <Tab label="A" value="a" />
     <Tab label="B" value="b" />
-  </Tabs></>
+  </Tabs>
+</>
 );