+++ codemods/misc/migrate-mui-chip-to-tag/CHANGELOG.md
@@ -1,1 +1,7 @@
 # @backstage/migrate-mui-chip-to-tag
+
+## 0.1.0
+
+### Minor Changes
+
+- Initial release: migrate MUI Chip to Backstage UI Tag (package omits -bui- because the BUI export is Tag).