+++ codemods/misc/migrate-mui-list-family-to-bui-list/CHANGELOG.md
@@ -1,1 +1,7 @@
 # @backstage/migrate-mui-list-family-to-bui-list
+
+## 0.1.0
+
+### Minor Changes
+
+- Initial release: migrate MUI List family rows to Backstage UI List / ListRow during the MUI 4 to BUI migration.