+++ codemods/misc/migrate-mui-tabs-to-bui-tabs/CHANGELOG.md
@@ -1,1 +1,7 @@
 # @backstage/migrate-mui-tabs-to-bui-tabs
+
+## 0.1.0
+
+### Minor Changes
+
+- Initial release: migrate MUI Tabs / TabContext patterns to Backstage UI Tabs during the MUI 4 to BUI migration.